requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
sqlalchemy>=2.0.9
pymysql>=1.0.3
python-multipart>=0.0.6
python-jose>=3.3.0
passlib>=1.7.4
requests>=2.28.2
pyjwt>=2.6.0
python-dotenv>=1.0.0
langchain-deepseek>=0.1.3
langchain-core>=0.3.54
langchain-openai>=0.3.9
openai>=1.75.0
tiktoken>=0.9.0
tenacity>=9.1.2